Remove invalid precision check
Check was removed becase it triggered false alarm.

Condider the following test case, that  caused exception to be incorrectly thrown:

    X = -1.09 * ?VOLT_DECIMALS_SCALESHIFT,
    X =/= trunc(X) %% is true

While X is definately fits into the BigDecimal, it gets corrupted (due to float nature) when expanded.
It is clear, that value _should_ still be written to the database despite of some (falsy) precision lost diagnostics